small request changes
