--- libs/movie/src/lib/movie/components/movie-filter/movie-filter.component.html
@@ -1,7 +1,0 @@
-<ng-container>
-  <h5>TITLE METADATA</h5>
-  <mat-expansion-panel *ngFor="let filter of filters">
-    <mat-expansion-panel-header>{{filter.label}}</mat-expansion-panel-header>
-    <ng-template [ngTemplateOutlet]="filter?.template"></ng-template>
-  </mat-expansion-panel>
-</ng-container>
