--- libs/movie/src/lib/movie/components/movie-filter/movie-filter.component.scss
@@ -1,19 +1,0 @@
-.filter-buttons {
-  margin: 24px 0;
-}
-
-mat-form-field {
-  width: 100%;
-}
-
-h5 {
-  margin-top: 20px;
-}
-
-mat-checkbox {
-  margin-top: 16px;
-}
-
-.checkbox-padding {
-  padding-left: 16px
-}
