+++ libs/movie/src/lib/movie/components/title-filter/title-filter.component.ts
@@ -1,4 +1,18 @@
-import { Component, ChangeDetectionStrategy } from '@angular/core';
+import {
+  Component,
+  ChangeDetectionStrategy,
+  Directive,
+  Input,
+  TemplateRef,
+  ContentChildren,
+  QueryList
+} from '@angular/core';
+
+@Directive({selector: '[filter]'})
+export class FilterDirective {
+  @Input() label = 'filter name';
+  constructor(public template: TemplateRef<any>) {}
+}
 
 @Component({
   selector: 'title-filter',
@@ -6,4 +20,6 @@ import { Component, ChangeDetectionStrategy } from '@angular/core';
   styleUrls: ['./title-filter.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class TitleFilterComponent {}
+export class TitleFilterComponent {
+  @ContentChildren(FilterDirective) filters: QueryList<FilterDirective>;
+}
