+++ libs/movie/src/lib/movie/components/title-filter/title-filter.module.ts
@@ -1,21 +1,23 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { TitleFilterComponent } from './title-filter.component';
+import { TitleFilterComponent, FilterDirective } from './title-filter.component';
 
 // Material
 import { MatCardModule } from '@angular/material/card';
 import { MatDividerModule } from '@angular/material/divider';
+import { MatExpansionModule } from '@angular/material/expansion';
 
 @NgModule({
-  declarations: [TitleFilterComponent],
+  declarations: [TitleFilterComponent, FilterDirective],
   imports: [
     CommonModule,
 
     // Material
     MatCardModule,
     MatDividerModule,
+    MatExpansionModule,
   ],
-  exports: [TitleFilterComponent]
+  exports: [TitleFilterComponent, FilterDirective]
 })
 export class TitleFilterModule {}
